WORKING STUFFgit add -Agit add -A
